+++ spec/rubocop/cops/style/comment_annotation_spec.rb
@@ -6,10 +6,23 @@ module Rubocop
   module Cop
     module Style
       describe CommentAnnotation do
+        before do
+          CommentAnnotation.config = {
+            'Keywords' => %w(TODO FIXME OPTIMIZE HACK REVIEW)
+          }
+        end
         let(:cop) { CommentAnnotation.new }
 
         it 'registers an offence for a missing colon' do
           inspect_source(cop, ['# TODO make better'])
+          expect(cop.offences).to have(1).item
+        end
+
+        it 'registers an offence for a missing colon after configured word' do
+          CommentAnnotation.config = {
+            'Keywords' => %w(ISSUE)
+          }
+          inspect_source(cop, ['# ISSUE wrong order'])
           expect(cop.offences).to have(1).item
         end
 
@@ -54,6 +67,21 @@ module Rubocop
 
         it 'accepts a comment that is obviously a code example' do
           inspect_source(cop, ['# Todo.destroy(1)'])
+          expect(cop.offences).to be_empty
+        end
+
+        it 'accepts a keyword that is just the beginning of a sentence' do
+          inspect_source(cop,
+                         ["# Optimize if you want. I wouldn't recommend it.",
+                          '# Hack is a fun game.'])
+          expect(cop.offences).to be_empty
+        end
+
+        it 'accepts a word without colon if not in the configuration' do
+          CommentAnnotation.config = {
+            'Keywords' => %w(FIXME OPTIMIZE HACK REVIEW)
+          }
+          inspect_source(cop, ['# TODO make better'])
           expect(cop.offences).to be_empty
         end
       end